+++ src/main/java/com/example/spacecourses/Services/IServicesQuiz.java
@@ -27,7 +27,7 @@ public interface IServicesQuiz {
 
     List<Result> getTopScore();
 
-    Integer getScore( Integer idU);
+    Double getScore( Integer idU);
